# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Structured report payload (see frontend renderer for full shape)
 */
export type DailyReportInputReportData = { [key: string]: unknown };

export interface DailyReportInput {
  /** ISO date YYYY-MM-DD */
  reportDate: string;
  /** Structured report payload (see frontend renderer for full shape) */
  reportData: DailyReportInputReportData;
}

export type DailyReportRecordReportData = { [key: string]: unknown };

export interface DailyReportRecord {
  id: number;
  reportDate: string;
  reportData: DailyReportRecordReportData;
  createdAt: string;
}