# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: daily-report
    description: Daily report operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /daily-report:
    post:
      operationId: submitDailyReport
      tags: [daily-report]
      summary: Submit daily report (called by N8N)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyReportInput"
      responses:
        "201":
          description: Report saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReportRecord"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /daily-report/latest:
    get:
      operationId: getLatestDailyReport
      tags: [daily-report]
      summary: Get latest daily report
      responses:
        "200":
          description: Latest report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReportRecord"
        "404":
          description: No report found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DailyReportInput:
      type: object
      required: [reportDate, reportData]
      properties:
        reportDate:
          type: string
          description: ISO date YYYY-MM-DD
        reportData:
          type: object
          description: Structured report payload (see frontend renderer for full shape)

    DailyReportRecord:
      type: object
      required: [id, reportDate, reportData, createdAt]
      properties:
        id:
          type: integer
        reportDate:
          type: string
        reportData:
          type: object
        createdAt:
          type: string