# lib/api-zod/src/generated/types/dailyReportInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyReportInputReportData } from "./dailyReportInputReportData";

export interface DailyReportInput {
  /** ISO date YYYY-MM-DD */
  reportDate: string;
  /** Structured report payload (see frontend renderer for full shape) */
  reportData: DailyReportInputReportData;
}