# lib/api-zod/src/generated/types/dailyReportRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyReportRecordReportData } from "./dailyReportRecordReportData";

export interface DailyReportRecord {
  id: number;
  reportDate: string;
  reportData: DailyReportRecordReportData;
  createdAt: string;
}